scripts/requirements-m.txt
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorzero==2.0
flatbuffers==20181003210633
gast==0.6.0
google-pasta==0.2.0
gpiozero==2.0.1
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.1.3
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
protobuf==5.29.4
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
